Fix issues based on editorconfig updates
