+++ sdk/storage/Azure.Storage.Blobs/tests/UnitTests.cs
@@ -1,4 +1,7 @@
-﻿using Azure.Storage.Blobs.Models;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Azure.Storage.Blobs.Models;
 using Azure.Storage.Common.Tests.Shared;
 using NUnit.Framework;
