+++ sdk/storage/Azure.Storage.Common/tests/Shared/MockStream.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
 using System.IO;
 
 namespace Azure.Storage.Common.Tests.Shared
